requirements.txt
# Formatters
black==25.1.0

# Linters
flake8==7.1.1
flake8-bugbear==24.12.12
flake8-pyproject==1.2.3
mypy==1.15.0
isort==6.0.1

# Testing frameworks
pytest==8.3.5
pytest-cov==6.1.0

# chess
python-chess==1.999
pygame-ce==2.5.6
